Add support for private registries when building images (#35)
When building Docker images, the FROM instruction can point to a
private Docker image. In order to pull a base image the registry may
require authentication. Similar to the create endpoint, there is a
`X-Registry-Config` header which luckily docker-modem already sets for
 us.

This PR simply adds an optional parameter to the method which allows
to forward the required configuration to docker-modem.